pyproject.toml
[project]
name = "pdf-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex",email = "aswann@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "img2pdf (>=0.5.1,<0.6.0)",
    "pypdf (>=5.2.0,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "filetype (>=1.2.0,<2.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "ruff (>=0.9.4,<0.10.0)",
    "python-docx (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pdf-tools = "pdf_tools.main:app"


[tool.mypy]
python_executable = "$VIRTUAL_ENV/bin/python"
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
show_error_codes = true
warn_unused_ignores = true
# TODO: enforce this in a future version
# warn_return_any = true
# disallow_any_unimported = true
exclude = '''(?x)(
    ^tests/
)'''


[tool.ruff]
target-version = "py311"
line-length = 79
indent-width = 4


[tool.ruff.lint]
select = ["B", "E", "F", "W", "C901", "ERA001"]
ignore = []
unfixable = ["B", "ERA001"]
extend-select = [
    "UP", "FIX002", "I"
]


[tool.ruff.lint.mccabe]
max-complexity = 10


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

